+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [#4222](https://github.com/influxdb/influxdb/pull/4222): Graphite TCP connections should not block shutdown
 - [#4180](https://github.com/influxdb/influxdb/pull/4180): Cursor & SelectMapper Refactor
 - [#1577](https://github.com/influxdb/influxdb/issues/1577): selectors (e.g. min, max, first, last) should have equivalents to return the actual point
+- [#4264](https://github.com/influxdb/influxdb/issues/4264): Refactor map functions to use list of values
 
 ## v0.9.4 [2015-09-14]